Add files via upload
ContainsLink will search files in a specified directory and its sub-directories, and return a list of links that are in the files. This program is helpful when you need to find and update links in many files - such as when updating documentation after moving to a new web address, network share, or SharePoint site.